reorganize into a class
